mysite/src/requirements.txt
Django==1.6.5
django-allauth==0.17.0
django-bootstrap-form==3.1
oauthlib==0.6.3
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
